chore!: eliminate dependency on the gaf
The go-application-framework should not be a required dependency of this
package. It provides only minimal supporting functionality:

- Convenience of access to a logger
- Convenience of two configuration options (Organization and IsFedramp)

While pulling in a lot of dependencies that a Snyk Code API client does not
really need. It should not be necessary, for example, to have to know about the GAF's
workflow engine in order to work with Snyk Code.

This change removes the dependency on the GAF, while leaving room to integrate
the code client with the GAF in the CLI and LS. This reduction of coupling
makes the open-source Snyk Code client more relevant to a much wider context of
reuse.